+++ src/ui/main/mod.rs
@@ -16,6 +16,7 @@ mod update_main_patch;
 mod download_wine;
 mod create_prefix;
 mod download_diff;
+mod disable_telemetry;
 mod launch;
 
 use anime_launcher_sdk::components::loader::ComponentsLoader;
@@ -66,9 +67,6 @@ pub enum AppMsg {
         /// Perform action when game or voice downloading is required
         /// Needed for chained executions (e.g. update one voice after another)
         perform_on_download_needed: bool,
-
-        /// Automatically start patch applying if possible and needed
-        apply_patch_if_needed: bool,
 
         /// Show status gathering progress page
         show_status_page: bool
@@ -306,6 +304,8 @@ impl SimpleComponent for App {
                                                 Some(LauncherState::PatchNotInstalled) |
                                                 Some(LauncherState::PatchUpdateAvailable) => "document-save-symbolic",
 
+                                                Some(LauncherState::TelemetryNotDisabled) => "security-high-symbolic",
+
                                                 Some(LauncherState::WineNotInstalled) |
                                                 Some(LauncherState::PrefixNotExists) => "document-save-symbolic",
 
@@ -330,6 +330,8 @@ impl SimpleComponent for App {
 
                                                 Some(LauncherState::PatchNotInstalled) |
                                                 Some(LauncherState::PatchUpdateAvailable) => tr("download-patch"),
+
+                                                Some(LauncherState::TelemetryNotDisabled) => tr("disable-telemetry"),
 
                                                 Some(LauncherState::PatchBroken) => tr("patch-broken"),
                                                 Some(LauncherState::PatchUnsafe) => tr("patch-unsafe"),
@@ -695,7 +697,6 @@ impl SimpleComponent for App {
             // Update launcher state
             sender.input(AppMsg::UpdateLauncherState {
                 perform_on_download_needed: false,
-                apply_patch_if_needed: false,
                 show_status_page: true
             });
 
@@ -715,7 +716,7 @@ impl SimpleComponent for App {
 
         match msg {
             // TODO: make function from this message like with toast
-            AppMsg::UpdateLauncherState { perform_on_download_needed, apply_patch_if_needed, show_status_page } => {
+            AppMsg::UpdateLauncherState { perform_on_download_needed, show_status_page } => {
                 if show_status_page {
                     sender.input(AppMsg::SetLoadingStatus(Some(Some(tr("loading-launcher-state")))));
                 } else {
@@ -823,6 +824,8 @@ impl SimpleComponent for App {
 
                     LauncherState::PatchNotInstalled |
                     LauncherState::PatchUpdateAvailable => update_main_patch::update_main_patch(sender, self.progress_bar.sender().to_owned()),
+
+                    LauncherState::TelemetryNotDisabled => disable_telemetry::disable_telemetry(sender),
 
                     LauncherState::WineNotInstalled => download_wine::download_wine(sender, self.progress_bar.sender().to_owned()),
                     LauncherState::PrefixNotExists => create_prefix::create_prefix(sender),